requirements.txt
google-api-python-client
youtube-transcript-api
pandas
scikit-learn
numpy
pyyaml
matplotlib